Changes addressing PR comments
